Move linters into seperate module. Give linters an abstract type to inherit, since they share an interface. Add empty README
