--- printer-linter/src/printerlinter/profile.py
@@ -1,7 +1,0 @@
-class Profile:
-    def __init__(self, file, settings) -> None:
-        self._settings = settings
-        self._file = file
-
-    def check(self) -> None:
-        yield
